Settings page - remove whitespace on top of page
I think just we can <umb-panel> or <merchello-panel>. <merchello-panel>
seems to be a replacement for <umb-panel> in some cases. When it is
wrapped inside the dashboard it seems to be pushed down (101px = height
of header). But if I don't use <umb-panel> or <merchello-panel> the
labels + descriptions are not aligned to the left. So I have just forced
it to be postioned to the top (it shouldn't be pushed down anyway inside
the dashboard).
I have also ensured the green button has a white text color.